[2018 spring rework] Implémentation de l'utilisation du WorldCollector dans world3D. Création du squelette de la classe WorldCollector. Implémentation du ChunkSystem.

+++ modules/worldapi/world/MapGenerator.cpp
@@ -20,9 +20,9 @@ ReliefMapGenerator::ReliefMapGenerator() {
 
 
 // -----
-const float CustomWorldRMGenerator::PIXEL_UNIT = 10;
+const double CustomWorldRMGenerator::PIXEL_UNIT = 10;
 
-CustomWorldRMGenerator::CustomWorldRMGenerator(float biomeDensity, uint32_t limitBrightness) :
+CustomWorldRMGenerator::CustomWorldRMGenerator(double biomeDensity, uint32_t limitBrightness) :
 	_biomeDensity(biomeDensity),
 	_limitBrightness(limitBrightness),
 	_diffLaw(std::make_unique<relief::CustomWorldDifferential>()) {
